ENT-6378 Take in a `Map` of args from reflective call
To allow Corda to init the shell with the right configuration, the class
`ShellConfiguration` has a `Map` representation on its public methods.

When called from Corda via reflection it constructs a `Map` and executes
the method. This `Map` is then turned into a `ShellConfiguration`
instance and used in the shell code.