ENT-6378 Copy ENT code over and apply some OS changes
Once the OS branch is complete I will merge is into the ENT branch and
copy code from this branch into a new one built on top of the OS
changes.

ENT-6378 Include parameter names compile flag

ENT-6378 Fix publish to maven local

ENT-6378 Take in a `Map` of args from reflective call

To allow Corda to init the shell with the right configuration, the class
`ShellConfiguration` has a `Map` representation on its public methods.

When called from Corda via reflection it constructs a `Map` and executes
the method. This `Map` is then turned into a `ShellConfiguration`
instance and used in the shell code.

ENT-6378 Fat jar :shell jar

Fat jar the jar so that when it is loaded from Corda that it has all the
dependencies that it needs.

Possibly means that we remove the `:shell-cli` module and turn it into a
single artifact.