ENT-6378 Migrate `corda-shell` to external repo (#1)
The `corda-shell` has been migrated out of the main Corda codebase and into this external repository.

`corda-tools-shell-cli` was removed and its code moved into `corda-shell`. This means there is now a single fat jar (`corda-shell`) which is used when run inside a node (by installing in a node's `/drivers` directory) or as a standalone jar. The only difference is entry point to the code inside the jar.

`corda-shell` (`:shell`) has had a new function added to `InteractiveShell` to allow it to be called reflectively from the Corda codebase. `startShell` takes in a `Map` of input configuration that is then converted to `ShellConfiguration` and then calls the existing version of `startShell`.

`InteractiveShell` was also modified further to add a `stop` method which tears down the shell which is useful for integration tests. `SSHServerTest` required this change as it was starting Corda nodes with SSH config, however the driver cannot start the SSH port since it requires the shell to do so, which was just moved out of Corda... Instead the test was modified to go through `InteractiveShell` directly and highlighted that the singleton class of `InteractiveShell` does not terminate cleanly (who'd have thought...).