ENT-6704 Separate standalone shell again
Split the standalone shell into its own module again.

This is so when the shell is run embedded in a Corda node, there won't
be 2 implementations of SLF4J.

The embedded `corda-shell` and `corda-standalone-shell` jar will do the
same functionality wise, just one can log on its own and one cannot.