ENT-6712: Invalidate cached RPC connection in case of connection loss
When non-`CordaRPCOps` is used for RPC communication, it does not re-connect automatically at connection loss. Therefore, it is prudent to effectively discard previously created instance of RPC connection and proxy such that a fresh instance will be created next time.

For `CordaRPCOps` no changes are made as it has `GracefulReconnect` mechanism to be resilient for connection loss.