ENT-6712: Invalidate cached RPC connection in case of connection loss (#25)
When non-`CordaRPCOps` is used for RPC communication, it does not re-connect automatically at connection loss. Therefore, it is prudent to effectively discard previously created instance of RPC connection and proxy such that a fresh instance will be created next time.

For `CordaRPCOps` no changes are made as it has `GracefulReconnect` mechanism to be resilient for connection loss.

Also eliminate all the references of `PartyInfoRpcOps` and replace with `CordaRPCOps`.
Since the RPC proxy for the latter has a more reliable re-connection mechanism.